Revert "Fix track not playable bug (#1220)"
This reverts commit 926d83888063b1151b6e7edf8855275735b33a6d.
