+++ packages/web/src/pages/track-page/store/sagas.js
@@ -1,8 +1,7 @@
 import { push as pushRoute } from 'connected-react-router'
 import moment from 'moment'
-import { call, fork, put, select, takeEvery } from 'redux-saga/effects'
+import { fork, call, put, select, takeEvery } from 'redux-saga/effects'
 
-import Kind from 'common/models/Kind'
 import { StringKeys } from 'common/services/remote-config'
 import * as trackCacheActions from 'common/store/cache/tracks/actions'
 import { getTrack as getCachedTrack } from 'common/store/cache/tracks/selectors'
@@ -12,13 +11,11 @@ import { getUsers } from 'common/store/cache/users/selectors'
 import * as trackPageActions from 'common/store/pages/track/actions'
 import { tracksActions } from 'common/store/pages/track/lineup/actions'
 import {
-  getSourceSelector,
   getTrack,
   getTrendingTrackRanks,
   getUser
 } from 'common/store/pages/track/selectors'
 import { getIsReachable } from 'common/store/reachability/selectors'
-import { makeUid } from 'common/utils/uid'
 import tracksSagas from 'pages/track-page/store/lineups/tracks/sagas'
 import apiClient from 'services/audius-api-client/AudiusAPIClient'
 import { remoteConfigInstance } from 'services/remote-config/remote-config-instance'
@@ -97,23 +94,11 @@ function* getTrackRanks(trackId) {
   yield put(trackPageActions.getTrackRanks(trackId))
 }
 
-function* addTrackToLineup(track) {
-  const source = yield select(getSourceSelector)
-  const formattedTrack = {
-    kind: Kind.TRACKS,
-    id: track.track_id,
-    uid: makeUid(Kind.TRACKS, track.track_id, source)
-  }
-
-  yield put(tracksActions.add(formattedTrack, track.track_id))
-}
-
-/** Get "more by this artist" and put into the lineup + queue */
-function* getRestOfLineup(permalink, ownerHandle) {
+function* getMoreByThisArtist(permalink, ownerHandle) {
   yield put(
-    tracksActions.fetchLineupMetadatas(1, 5, false, {
+    tracksActions.fetchLineupMetadatas(0, 6, false, {
       ownerHandle,
-      heroTrackPermalink: permalink
+      permalink
     })
   )
 }
@@ -122,6 +107,7 @@ function* watchFetchTrack() {
   yield takeEvery(trackPageActions.FETCH_TRACK, function* (action) {
     const { trackId, handle, slug, canBeUnlisted } = action
     const permalink = `/${handle}/${slug}`
+    yield fork(getMoreByThisArtist, permalink, handle)
     try {
       let track
       if (!trackId) {
@@ -152,11 +138,6 @@ function* watchFetchTrack() {
           return
         }
       } else {
-        yield put(trackPageActions.setTrackId(track.track_id))
-        // Add hero track to lineup early so that we can play it ASAP
-        // (instead of waiting for the entire lineup to load)
-        yield call(addTrackToLineup, track)
-        yield fork(getRestOfLineup, permalink, handle)
         yield fork(getTrackRanks, track.track_id)
         yield put(trackPageActions.fetchTrackSucceeded(track.track_id))
       }
@@ -188,12 +169,7 @@ function* watchRefetchLineup() {
     const { permalink } = yield select(getTrack)
     const { handle } = yield select(getUser)
     yield put(tracksActions.reset())
-    yield put(
-      tracksActions.fetchLineupMetadatas(0, 6, false, {
-        ownerHandle: handle,
-        heroTrackPermalink: permalink
-      })
-    )
+    yield call(getMoreByThisArtist, permalink, handle)
   })
 }
